+++ packages/talaria/lib/talaria.dart
@@ -20,4 +20,5 @@ export 'src/tracing/trace_context.dart';
 export 'src/tracing/tracer.dart' show Tracer, SpanEnrichment;
 export 'src/integration/zone_integration.dart' show runZonedTalaria;
 export 'src/transport/http_transport.dart' show HttpTransport;
+export 'src/transport/ingest_error.dart' show IngestError, IngestSignal;
 export 'src/transport/transport.dart' show Transport, TransportException;